[수정] 사용자의 위치정보를 가져오는 LocationService 생성
