+++ OneDay/OneDay/Controller/EntryViewController.swift
@@ -9,7 +9,6 @@
 import UIKit
 import CoreData
 import MobileCoreServices
-import CoreLocation
 
 class EntryViewController: UIViewController {
     
@@ -30,8 +29,6 @@ class EntryViewController: UIViewController {
     
     lazy var isImageSelected = false
     
-    let locationManager = CLLocationManager()
-    
     // MARK: - Life cycle
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -46,15 +43,7 @@ class EntryViewController: UIViewController {
         timeLabel.text = dateSet.time
         textView.textDragDelegate = self
         
-        // 사용자 위치정보 권한 요청
-        locationManager.requestWhenInUseAuthorization()
-        
-        if CLLocationManager.locationServicesEnabled() {
-            locationManager.delegate = self
-            locationManager.desiredAccuracy = kCLLocationAccuracyNearestTenMeters
-            locationManager.startUpdatingLocation()
-        }
-        
+        setUpWeather()
         setUpPreview()
     }
     
@@ -204,13 +193,5 @@ extension EntryViewController: UITextDragDelegate {
         } else {
             return []
         }
-    }
-}
-
-extension EntryViewController: CLLocationManagerDelegate {
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        guard let localValue: CLLocationCoordinate2D = manager.location?.coordinate else { return }
-        
-        loadWeatherInfomation(latitude: "\(localValue.latitude)", longitude: "\(localValue.longitude)")
     }
 }
